more porting progress - Networking next

+++ src/main/java/net/Pandarix/betterarcheology/mixin/FasterBrushingMixin.java
@@ -21,7 +21,7 @@ public abstract class FasterBrushingMixin
     {
         if (player.getActiveItem().getItem() instanceof BetterBrushItem brushItem)
         {
-            this.nextBrushTime -= (10L - brushItem.getBrushingSpeed());
+            this.nextBrushTime -= (long) (10L - brushItem.getBrushingSpeed());
         }
     }
 }
